requirements.txt
pandas==2.1.4 
numpy==1.26.2 
pyspark==3.5.0 
mysql-connector-python==8.3.0 
SQLAlchemy==2.0.25 
python-dotenv==1.0.0 
pymongo==4.6.1

